Heloworld/requirements.txt
asn1crypto==1.4.0
attrs==21.2.0
azure-common==1.1.27
azure-core==1.17.0
azure-storage-blob==12.8.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bleach==4.0.0
boto3==1.18.28
botocore==1.21.28
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
configparser==5.0.2
cryptography==3.4.8
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.2
docutils==0.17.1
entrypoints==0.3
filelock==3.0.12
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
gunicorn==20.1.0
idna==2.8
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isodate==0.6.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.1
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-Garden==0.1.4
lockfile==0.12.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
msrest==0.6.21
mysql-connector-python==8.0.26
mysqlclient==2.0.3
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
numpy==1.21.2
oauthlib==3.1.1
opencv-python==4.5.3.56
oscrypto==1.2.1
packaging==21.0
pandas==1.3.2
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
pipenv==2021.5.29
platformdirs==2.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.17.3
psycopg2==2.9.1
pycparser==2.20
pycryptodomex==3.10.1
Pygments==2.10.0
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
pysnowflake==0.1.3
python-daemon==2.2.4
python-dateutil==2.8.2
pytz==2021.1
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.10.0
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
Send2Trash==1.8.0
six==1.16.0
smtpapi==0.4.7
snowflake-connector-python==2.5.1
snowflake-python-sdk==1.0.0
snowflake-sqlalchemy==1.3.1
SQLAlchemy==1.4.23
terminado==0.11.1
testpath==0.5.0
tk==0.1.0
tornado==6.1
traitlets==5.0.5
urllib3==1.25.11
virtualenv==20.7.2
virtualenv-clone==0.5.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
